requirements.txt
streamlit
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
Pillow
